[EI-125] [api] rename: persistent -> persistence
