+++ api/src/community/community.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { CqrsModule } from '@nestjs/cqrs';
 import { CommunityController } from './api/community.controller';
 import { CreatePostCommandHandler } from './application/command/create-post/create-post.command.handler';
-import { CommunityPersistentAdapter } from './infrastructure/adapter/persistent/community.persistent.adapter.stub';
+import { CommunityPersistentAdapter } from './infrastructure/adapter/persistence/community.persistent.adapter.stub';
 import { GetPostsQueryHandler } from './application/query/get-posts/get-posts.query.handler';
 
 @Module({
