+++ api/src/numerical-guidance/infrastructure/adapter/persistence/entity/indicator-board-meta-data.entity.ts
