+++ api/src/numerical-guidance/infrastructure/adapter/persistence/indicator-board-meta-data.persistent.adapter.ts
