+++ api/src/numerical-guidance/infrastructure/adapter/persistence/mapper/indicator-board-meta-data.mapper.ts
